simple_rag/requirements.txt
streamlit
fastapi
uvicorn
pypdf
faiss-cpu
langchain
langchain-openai
neo4j
